Restructure in separate files
